Update app.js
Removed localhost from the fetch address, because it was trying to access the local host. but on deploying on a platform, it should access their local host, not mine. So i declared a env variable for port and put it in app.listen() in src/app.js , and here removed the localhost:4000 from the fetch address